CAMEL-16488: ensure resource setup/cleanup on the right time
